feat(SaveLoadService.cs): add SaveWorldMapAsync method to ISaveLoadService interface
chore(Vega.Framework.csproj): add BinaryPack package reference

fix(WorldMap.cs): remove unnecessary empty line

The changes were made to add a new method `SaveWorldMapAsync` to the `ISaveLoadService` interface. This method allows saving a `WorldMap` object to a specified path.

Additionally, the `BinaryPack` package was added as a reference to the `Vega.Framework.csproj` file to support serialization and deserialization of objects.

In the `WorldMap.cs` file, an unnecessary empty line was removed for code cleanliness.