feat(MapSpawnerService.cs): add method SearchLandTileByFlag to search for land tiles by flag fix(WorldService.cs): change return type of PlaceLands method to IEnumerable<LandGameObject> to return the generated land game objects feat(WorldService.cs): add method PlacePlayer to place the player on a random land tile feat(WorldPlayerGameObject.cs): create a new class for the player game object in the world map
The changes were made to enhance the functionality of the map spawner service and the world service. The MapSpawnerService now includes a new method called SearchLandTileByFlag, which allows searching for land tiles by a specific flag. This method will be used to find the player tile. The WorldService has been modified to change the return type of the PlaceLands method to IEnumerable<LandGameObject> to return the generated land game objects. Additionally, a new method called PlacePlayer has been added to place the player on a random land tile. Finally, a new class called WorldPlayerGameObject has been created to represent the player game object in the world map.